Move Macro to abstract/macro.py
