+++ myia/abstract/__init__.py
@@ -4,5 +4,6 @@ from .aliasing import *  # noqa
 from .data import *  # noqa
 from .infer import *  # noqa
 from .loop import *  # noqa
+from .macro import *  # noqa
 from .ref import *  # noqa
 from .utils import *  # noqa
